requirements.txt

beautifulsoup4
chromadb
crewai
crewai-tools
google-auth
ipykernel
langchain
langchain-openai
openai
pandas
redis
redisvl
selenium==4.34.0
streamlit
tiktoken
python-dotenv
sentence_transformers
streamlit-oauth
requests
tokenizers==0.20.3
webdriver-manager==4.0.2
tavily-python